NETVIRT 1202 : AS number issue during display Value of AS number was an integer which gave rise to negative values so changing it to long to increase the range of the value Signed-off-by: Loshmitha <loshmitha@ericsson.com>

+++ bgpmanager/impl/src/main/java/org/opendaylight/netvirt/bgpmanager/commands/Cache.java
@@ -149,7 +149,7 @@ public class Cache extends OsgiCommandSupport {
         }
         AsId asId = config.getAsId();
         if (asId != null) {
-            int asNum = asId.getLocalAs().intValue();
+            long asNum = asId.getLocalAs().longValue();
             IpAddress routerId = asId.getRouterId();
             Long spt = asId.getStalepathTime();
             Boolean afb = asId.isAnnounceFbit();
